Add files via upload
With the rapid development of Internet technology, online applications have become an indispensable part of government, enterprise and engineering fields. Most enterprises have utilized the advantages of cloud computing and refactored the applications with the service-oriented architecture (SOA) in order to solve the huge number of online users, soaring data and fluctuant load. According to the cloud services quality which evaluated in different cloud computing resources, developer choose the suitable size of the service cluster to meet the needs of enterprise services in the distributed service cluster. It can not only ensure the stable operation of online applications, but also maximize the reduction of basic costs. Therefore, it is important to evaluate the quality of cloud services accurately. This paper proposes an evaluation method of cloud service quality based on queuing model. The M/M/1 and M/M/s queuing models are used to analyze and evaluate the quality of single node service and cluster service respectively. By the method proposed in this paper can not only choose the suitable size of the service cluster for online applications, but also perform dynamic scheduling of cloud resources. Finally, the feasibility and practicability of this method is verified by experiments.